requirements.txt
pymysql
pyodbc
flask
flask_restful
flask_sqlalchemy
lxml=4.9.3
pyOpenSSL
pycryptodome   
pydantic
signxml=2.5.2